+++ rio/ChangeLog.md
@@ -3,6 +3,7 @@
 ## 0.1.3.0
 
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
+* Add `foldMapM`
 
 ## 0.1.2.0