+++ rio/ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## 0.1.3.0
 
+* Add `newLogFunc` function to create `LogFunc` records outside of a callback scope
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
 
 ## 0.1.2.0