+++ rio/ChangeLog.md
@@ -4,6 +4,9 @@
 
 * Add `newLogFunc` function to create `LogFunc` records outside of a callback scope
 * Allow dynamic reloading of `logMinLevel` and `logVerboseFormat` for the `LogOptions` record
+* Add `headMaybe`, `lastMaybe`, `tailMaybe`, `initMaybe`, `maximumMaybe`, `minimumMaybe`,
+  `maximumByMaybe`, `minimumByMaybe` functions to `RIO.List` module (issue #82)
+* Move non partial functions `scanr1` and `scanl1` from `RIO.List.Partial` to `RIO.List` (issue #82)
 
 ## 0.1.2.0